TEP-IFPI-2017_8-master/django-api-atv1-master/requirements.txt
certifi==2018.11.29
chardet==3.0.4
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.3
django-debug-toolbar==1.10.1
django-fake-model==0.1.4
django-import-export==1.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
drf-extensions==0.4.0
drf-nested-routers==0.91
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0b10
httplib2==0.12.0
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.12
Pillow==5.3.0
psycopg2==2.7.6.1
PyPDF2==1.26.0
python-decouple==3.1
python-firebase==1.2
pytz==2018.7
PyYAML==3.13
raven==6.9.0
reportlab==3.5.12
requests==2.20.1
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1.2
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0
